chore: move shared fixtures into conftest.py
