+++ samples/snippets/create_data_labeling_job_active_learning_sample_test.py
@@ -18,16 +18,15 @@ import uuid
 from google.cloud import aiplatform
 import pytest
 
-import create_data_labeling_job_specialist_pool_sample
+import create_data_labeling_job_active_learning_sample
 import helpers
 
 API_ENDPOINT = os.getenv("DATA_LABELING_API_ENDPOINT")
 PROJECT_ID = os.getenv("BUILD_SPECIFIC_GCLOUD_PROJECT")
 LOCATION = "us-central1"
 DATASET_ID = "1905673553261363200"
-SPECIALIST_POOL_ID = "5898026661995085824"
 INPUTS_SCHEMA_URI = "gs://google-cloud-aiplatform/schema/datalabelingjob/inputs/image_classification_1.0.0.yaml"
-DISPLAY_NAME = f"temp_create_data_labeling_job_specialist_pool_test_{uuid.uuid4()}"
+DISPLAY_NAME = f"temp_create_data_labeling_job_active_learning_test_{uuid.uuid4()}"
 
 INSTRUCTIONS_GCS_URI = (
     "gs://ucaip-sample-resources/images/datalabeling_instructions.pdf"
@@ -35,35 +34,24 @@ INSTRUCTIONS_GCS_URI = (
 ANNOTATION_SPEC = "rose"
 
 
-@pytest.fixture
-def shared_state():
-    state = {}
-    yield state
-
-
-@pytest.fixture
-def job_client():
-    client_options = {"api_endpoint": API_ENDPOINT}
-    job_client = aiplatform.gapic.JobServiceClient(client_options=client_options)
-    yield job_client
-
-
 @pytest.fixture(scope="function", autouse=True)
-def teardown(capsys, shared_state, job_client):
+def teardown(capsys, shared_state, data_labeling_job_client):
     yield
 
-    job_client.cancel_data_labeling_job(name=shared_state["data_labeling_job_name"])
+    data_labeling_job_client.cancel_data_labeling_job(
+        name=shared_state["data_labeling_job_name"]
+    )
 
     # Verify Data Labelling Job is cancelled, or timeout after 400 seconds
     helpers.wait_for_job_state(
-        get_job_method=job_client.get_data_labeling_job,
+        get_job_method=data_labeling_job_client.get_data_labeling_job,
         name=shared_state["data_labeling_job_name"],
         timeout=400,
         freq=10,
     )
 
     # Delete the data labeling job
-    response = job_client.delete_data_labeling_job(
+    response = data_labeling_job_client.delete_data_labeling_job(
         name=shared_state["data_labeling_job_name"]
     )
 
@@ -76,16 +64,12 @@ def teardown(capsys, shared_state, job_client):
 
 
 # Creating a data labeling job for images
-def test_create_data_labeling_job_specialist_pool_sample(capsys, shared_state):
+def test_create_data_labeling_job_active_learning_sample(capsys, shared_state):
 
-    dataset = f"projects/{PROJECT_ID}/locations/{LOCATION}/datasets/{DATASET_ID}"
-    specialist_pool = f"projects/{PROJECT_ID}/locations/{LOCATION}/specialistPools/{SPECIALIST_POOL_ID}"
-
-    create_data_labeling_job_specialist_pool_sample.create_data_labeling_job_specialist_pool_sample(
+    create_data_labeling_job_active_learning_sample.create_data_labeling_job_active_learning_sample(
         project=PROJECT_ID,
         display_name=DISPLAY_NAME,
-        dataset=dataset,
-        specialist_pool=specialist_pool,
+        dataset=f"projects/{PROJECT_ID}/locations/{LOCATION}/datasets/{DATASET_ID}",
         instruction_uri=INSTRUCTIONS_GCS_URI,
         inputs_schema_uri=INPUTS_SCHEMA_URI,
         annotation_spec=ANNOTATION_SPEC,
