+++ samples/snippets/create_training_pipeline_image_object_detection_sample_test.py
@@ -28,14 +28,8 @@ DATASET_ID = "3555732643297361920"  # permanent_salad_object_detection_dataset
 DISPLAY_NAME = f"temp_create_training_pipeline_image_obj_detection_test_{uuid4()}"
 
 
-@pytest.fixture
-def shared_state():
-    state = {}
-    yield state
-
-
 @pytest.fixture(scope="function", autouse=True)
-def teardown(shared_state):
+def teardown(shared_state, pipeline_client):
     yield
 
     training_pipeline_id = shared_state["training_pipeline_name"].split("/")[-1]
@@ -43,10 +37,6 @@ def teardown(shared_state):
     # Stop the training pipeline
     cancel_training_pipeline_sample.cancel_training_pipeline_sample(
         project=PROJECT_ID, training_pipeline_id=training_pipeline_id
-    )
-
-    pipeline_client = aiplatform.gapic.PipelineServiceClient(
-        client_options={"api_endpoint": "us-central1-aiplatform.googleapis.com"}
     )
 
     # Waiting for training pipeline to be in CANCELLED state
