+++ samples/snippets/create_training_pipeline_tabular_classification_sample_test.py
@@ -30,14 +30,8 @@ TARGET_COLUMN = "species"
 PREDICTION_TYPE = "classification"
 
 
-@pytest.fixture
-def shared_state():
-    state = {}
-    yield state
-
-
 @pytest.fixture(scope="function", autouse=True)
-def teardown(shared_state):
+def teardown(shared_state, pipeline_client):
     yield
 
     training_pipeline_id = shared_state["training_pipeline_name"].split("/")[-1]
@@ -45,11 +39,6 @@ def teardown(shared_state):
     # Stop the training pipeline
     cancel_training_pipeline_sample.cancel_training_pipeline_sample(
         project=PROJECT_ID, training_pipeline_id=training_pipeline_id
-    )
-
-    client_options = {"api_endpoint": "us-central1-aiplatform.googleapis.com"}
-    pipeline_client = aiplatform.gapic.PipelineServiceClient(
-        client_options=client_options
     )
 
     # Waiting for training pipeline to be in CANCELLED state
