+++ samples/snippets/create_training_pipeline_video_action_recognition_sample_test.py
@@ -31,22 +31,6 @@ DISPLAY_NAME = (
 )
 MODEL_DISPLAY_NAME = f"Temp Model for {DISPLAY_NAME}"
 MODEL_TYPE = "CLOUD"
-API_ENDPOINT = "us-central1-aiplatform.googleapis.com"
-
-
-@pytest.fixture
-def shared_state():
-    state = {}
-    yield state
-
-
-@pytest.fixture
-def pipeline_client():
-    client_options = {"api_endpoint": API_ENDPOINT}
-    pipeline_client = aiplatform.gapic.PipelineServiceClient(
-        client_options=client_options
-    )
-    yield pipeline_client
 
 
 @pytest.fixture(scope="function", autouse=True)
@@ -71,9 +55,7 @@ def teardown(shared_state, pipeline_client):
 
 
 # Training AutoML Vision Model
-def test_create_training_pipeline_video_action_recognition_sample(
-    capsys, shared_state
-):
+def test_create_training_pipeline_video_action_recognition_sample(capsys, shared_state):
     create_training_pipeline_video_action_recognition_sample.create_training_pipeline_video_action_recognition_sample(
         project=PROJECT_ID,
         display_name=DISPLAY_NAME,
