+++ samples/snippets/get_model_evaluation_video_action_recognition_sample_test.py
